Added better security, did some code cleanup, and improved the readme.

+++ featured-content.php
@@ -1,18 +1,18 @@
 <?php
 /*
- Plugin Name: Featured Content
- Description: Easily mark any content on your site featured with a simple checkbox or link.
- Version: 1.0.0
- Author: Nick Ohrn of Plugin-Developer.com
- Author URI: http://plugin-developer.com/
- */
+Plugin Name: Easy Featured Content
+Description: Easily mark any content on your site featured with a simple checkbox or link.
+Version: 1.1.0
+Author: Nick Ohrn of Plugin-Developer.com
+Author URI: http://plugin-developer.com/
+*/
 
 if(!class_exists('Featured_Content')) {
 	class Featured_Content {
 		/// CONSTANTS
 
 		//// VERSION
-		const VERSION = '1.0.0';
+		const VERSION = '1.1.0';
 
 		//// KEYS
 		const IS_FEATURED_CONTENT_KEY = '_is_featured_content';
@@ -34,14 +34,12 @@ if(!class_exists('Featured_Content')) {
 				add_action('admin_enqueue_scripts', array(__CLASS__, 'enqueue_administrative_resources'));
 			}
 
+			add_action('after_setup_theme', array(__CLASS__, 'add_post_type_support'), 9);
 			add_action('save_post', array(__CLASS__, 'save_post_meta'), 10, 2);
 			add_action('wp_ajax_featured-content', array(__CLASS__, 'ajax_featured_content_toggle'));
 		}
 
 		private static function add_filters() {
-			// Because we're allowing the toggle to happen via JavaScript, I don't want to mess with this right now
-			// add_filter('display_post_states', array(__CLASS__, 'add_featured_state'));
-
 			add_filter('pre_get_posts', array(__CLASS__, 'transform_query_variables'));
 
 			add_filter('page_row_actions', array(__CLASS__, 'add_featured_content_toggle'), 10, 2);
@@ -54,17 +52,28 @@ if(!class_exists('Featured_Content')) {
 			$data = stripslashes_deep($_REQUEST);
 
 			$is_ajax = isset($data['is-ajax']);
+
 			$is_featured_content = isset($data['is-featured-content']) && 'yes' === $data['is-featured-content'] ? 'yes' : 'no';
 			$is_featured_content_bool = 'yes' === $is_featured_content;
-			$post_id = isset($data['post-id']) ? $data['post-id'] : 0;
 
-			if($post_id) {
+			$nonce = isset($data['featured-content-toggle-nonce']) ? $data['featured-content-toggle-nonce'] : false;
+
+			$post_id = isset($data['post-id']) ? $data['post-id'] : 0;
+			$post = get_post($post_id);
+			$post_type_object = get_post_type_object($post->post_type);
+
+			$can_edit_post = current_user_can($post_type_object->cap->edit_post, $post->ID);
+			$supports_featured_content = post_type_supports($post->post_type, 'featured-content');
+
+			if($can_edit_post && $supports_featured_content && wp_verify_nonce($nonce, 'featured-content-toggle')) {
 				self::_set_is_featured_content($post_id, $is_featured_content);
 			}
 
+			$is_featured_content_bool = self::is_featured_content($post_id);
+
 			if($is_ajax) {
 				$results = array(
-					'link' => add_query_arg(array('action' => 'featured-content', 'is-featured-content' => $is_featured_content_bool ? 'no' : 'yes', 'post-id' => $post_id), admin_url('admin-ajax.php')),
+					'link' => self::_get_ajax_toggle_link($post_id, $is_featured_content_bool),
 					'text' => $is_featured_content_bool ? __('Unfeature') : __('Feature')
 				);
 
@@ -76,13 +85,25 @@ if(!class_exists('Featured_Content')) {
 			exit;
 		}
 
+		private static function _get_ajax_toggle_link($post_id, $is_featured_content_bool) {
+			return wp_nonce_url(add_query_arg(array(
+				'action' => 'featured-content',
+				'is-featured-content' => $is_featured_content_bool ? 'no' : 'yes',
+				'post-id' => $post_id
+			), admin_url('admin-ajax.php')), 'featured-content-toggle', 'featured-content-toggle-nonce');
+		}
+
 		/// CALLBACKS
 
 		public static function add_featured_content_toggle($actions, $post) {
 			$post_type_object = get_post_type_object($post->post_type);
-			if (current_user_can($post_type_object->cap->edit_post, $post->ID)) {
+
+			$can_edit_post = current_user_can($post_type_object->cap->edit_post, $post->ID);
+			$supports_featured_content = post_type_supports($post->post_type, 'featured-content');
+
+			if($can_edit_post && $supports_featured_content) {
 				$is_featured_content = self::is_featured_content($post->ID);
-				$link = add_query_arg(array('action' => 'featured-content', 'is-featured-content' => $is_featured_content ? 'no' : 'yes', 'post-id' => $post->ID), admin_url('admin-ajax.php'));
+				$link = self::_get_ajax_toggle_link($post->ID, $is_featured_content);
 				$text = $is_featured_content ? __('Unfeature') : __('Feature');
 
 				$actions['featured-content'] = sprintf('<a class="is-featured-content-toggle" href="%s">%s</a>', $link, $text);
@@ -91,18 +112,15 @@ if(!class_exists('Featured_Content')) {
 			return $actions;
 		}
 
-		public static function add_featured_state($post_states) {
-			$post_id = get_the_ID();
-
-			if(self::is_featured_content($post_id)) {
-				$post_states['is-featured-content'] = __('Featured');
+		public static function add_meta_boxes($post_type) {
+			if(post_type_supports($post_type, 'featured-content')) {
+				add_meta_box('featured-content-meta-box', __('Featured'), array(__CLASS__, 'display_meta_box'), $post_type, 'side', 'core');
 			}
-
-			return $post_states;
 		}
 
-		public static function add_meta_boxes($post_type) {
-			add_meta_box('featured-content-meta-box', __('Featured'), array(__CLASS__, 'display_meta_box'), $post_type, 'side', 'core');
+		public static function add_post_type_support() {
+			add_post_type_support('page', 'featured-content');
+			add_post_type_support('post', 'featured-content');
 		}
 
 		public static function enqueue_administrative_resources($hook) {
@@ -119,6 +137,7 @@ if(!class_exists('Featured_Content')) {
 			$data = stripslashes_deep($_POST);
 			if(wp_is_post_autosave($post_id)
 				|| wp_is_post_revision($post_id)
+				|| !post_type_supports($post->post_type, 'featured-content')
 				|| !isset($data['featured-content-save-meta-nonce'])
 				|| !wp_verify_nonce($data['featured-content-save-meta-nonce'], 'featured-content-save-meta')) {
 				return;
@@ -159,28 +178,39 @@ if(!class_exists('Featured_Content')) {
 		/// POST META
 
 		private static function _get_is_featured_content($post_id) {
-			$post_id = empty($post_id) ? get_the_ID() : $post_id;
-			$is_featured_content = wp_cache_get(self::IS_FEATURED_CONTENT_KEY, $post_id);
+			$post_id = empty($post_id) && in_the_loop() ? get_the_ID() : $post_id;
 
-			if(false === $is_featured_content) {
-				$is_featured_content = 'yes' === get_post_meta($post_id, self::IS_FEATURED_CONTENT_KEY, true) ? 'yes' : 'no';
-				wp_cache_set(self::IS_FEATURED_CONTENT_KEY, $is_featured_content, $post_id, time() + self::CACHE_PERIOD);
+			if($post_id) {
+				$is_featured_content = wp_cache_get(self::IS_FEATURED_CONTENT_KEY, $post_id);
+
+				if(false === $is_featured_content) {
+					$is_featured_content = 'yes' === get_post_meta($post_id, self::IS_FEATURED_CONTENT_KEY, true) ? 'yes' : 'no';
+					wp_cache_set(self::IS_FEATURED_CONTENT_KEY, $is_featured_content, $post_id, time() + self::CACHE_PERIOD);
+				}
+			} else {
+				$is_featured_content = null;
 			}
+
 
 			return 'yes' === $is_featured_content;
 		}
 
 		private static function _set_is_featured_content($post_id, $is_featured_content) {
-			$post_id = empty($post_id) ? get_the_ID() : $post_id;
-			$is_featured_content = 'yes' === $is_featured_content ? 'yes' : 'no';
+			$post_id = empty($post_id) && in_the_loop() ? get_the_ID() : $post_id;
 
-			if('yes' === $is_featured_content) {
-				update_post_meta($post_id, self::IS_FEATURED_CONTENT_KEY, $is_featured_content);
+			if($post_id) {
+				$is_featured_content = 'yes' === $is_featured_content ? 'yes' : 'no';
+
+				if('yes' === $is_featured_content) {
+					update_post_meta($post_id, self::IS_FEATURED_CONTENT_KEY, $is_featured_content);
+				} else {
+					delete_post_meta($post_id, self::IS_FEATURED_CONTENT_KEY);
+				}
+
+				wp_cache_delete(self::IS_FEATURED_CONTENT_KEY, $post_id);
 			} else {
-				delete_post_meta($post_id, self::IS_FEATURED_CONTENT_KEY);
+				$is_featured_content = false;
 			}
-
-			wp_cache_set(self::IS_FEATURED_CONTENT_KEY, $is_featured_content, $post_id, time() + self::CACHE_PERIOD);
 
 			return $is_featured_content;
 		}
@@ -190,8 +220,6 @@ if(!class_exists('Featured_Content')) {
 		public static function is_featured_content($post_id) {
 			return self::_get_is_featured_content($post_id);
 		}
-
-
 	}
 
 	require_once('lib/template-tags.php');
